requirements.txt
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==4.9.3
huggingface_hub>=0.34.0
